#60: Added separate .md file for About page
