+++ package.json
@@ -19,6 +19,7 @@
         "react-dom": "^16.13.0",
         "react-dropzone": "^10.2.2",
         "react-google-charts": "^3.0.15",
+        "react-md-file": "^2.0.0",
         "react-pdf": "^4.1.0",
         "react-redux": "^7.2.0",
         "react-router-dom": "^5.1.2",
