--- src/components/layouts/pages/about/About.scss
@@ -1,16 +1,0 @@
-@import "../../../variables";
-
-.service-list {
-  list-style: none;
-  padding: 0;
-
-  .service-info {
-    &__status {
-      margin-left: 10px;
-
-      &._error {
-        color: $color-error;
-      }
-    }
-  }
-}
